state estimation refactored
