+++ ros2_ws/src/algorithms/state_estimation/state_estimation_2d/setup.py
@@ -13,6 +13,7 @@ setup(
         ['resource/' + package_name]),
     ('share/' + package_name, ['package.xml']),
     (os.path.join('share', package_name, 'launch'), glob('launch/*.launch.py')),
+    (os.path.join('share', package_name, 'rviz'), glob('rviz/*.rviz'))
     ],
     install_requires=['setuptools'],
     zip_safe=True,
